Add forseti module templates for building AppEngine pipeline
